backend/requirements.txt
flask==3.0.2
pdfplumber==0.10.3
python-dotenv==1.0.0
google-generativeai==0.3.2
flask-cors==4.0.0
